Validations

+++ db/migrate/20200713085700_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[6.0]
     create_table :reviews do |t|
       t.string :title
       t.string :body
-      t.decimal :rating
+      t.integer :rating
       t.references :user, null: false, foreign_key: true
       t.references :deck, null: false, foreign_key: true
